TechdocsAPI/requirements.txt
fastapi==0.99.1
uvicorn
requests
pydantic==1.10.12
python-jose[cryptography]
passlib[bcrypt]
mysql-connector-python
langchain-google-genai
pydantic[email]
langchain
Pillow
jinja2
